feat: move kubernetes_cleanup_pods's config to top level script
